check if ec_sys file interface already exists before modprobe
The stock Arch Linux kernel has the 'ec_sys' module compiled in, which
means there is nothing to modprobe, causing an error. Make the code
check for the existence of this file interface before attempting to
modprobe.

Also, add a write check as the file interface is not writable by
default.