Add power reset button
Pressing the power reset button enables auto-suspend at 30 minutes
for all power profiles, and sets Screen blank to 5 minutes. Both
changes are for all active users.

This change is required by California Title 20 power regulations.
According to that regulation all System76 hardware must ship with
this feature to be legally sold in California (And states with
similar legislation.) This change is both retroactive to existing
sold systems, and future systems.